ファイル移動/機能完成
- サンプルプロジェクト追加のため, ファイルをフォルダに移動
- 機能を完成させ, デバッグ用ログ出力機能を除去
